Add new Method -- not working correctly

+++ src/Wifi.WeatherForeCast/Wifi.WeatherForeCast.GeodataApi/HelperJsonObject/GeoDataApiJsonModel.cs
@@ -2,10 +2,10 @@
 
 public class GeoDataApiJsonModel
 {
-    public int place_id { get; set; }
+    public Int64 place_id { get; set; }
     public string licence { get; set; }
     public string osm_type { get; set; }
-    public int osm_id { get; set; }
+    public Int64 osm_id { get; set; }
     public List<string> boundingbox { get; set; }
     public double lat { get; set; }
     public double lon { get; set; }
